# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Book:
    properties:
      author:
        type: string
      categories:
        items:
          $ref: '#/definitions/models.Category'
        type: array
      copies_count:
        minimum: 0
        type: integer
      cover_url:
        type: string
      created_at:
        type: string
      description:
        type: string
      download_count:
        type: integer
      files:
        items:
          $ref: '#/definitions/models.BookFile'
        type: array
      id:
        type: string
      is_premium:
        type: boolean
      isbn:
        type: string
      language:
        type: string
      page_count:
        type: integer
      publication_year:
        maximum: 9999
        minimum: 0
        type: integer
      publisher:
        type: string
      rating:
        type: number
      rating_count:
        type: integer
      status:
        $ref: '#/definitions/models.BookStatus'
      title:
        type: string
      updated_at:
        type: string
      view_count:
        type: integer
    required:
    - author
    - title
    type: object
  models.BookFile:
    properties:
      book:
        $ref: '#/definitions/models.Book'
      book_id:
        type: string
      created_at:
        type: string
      file_name:
        type: string
      file_size:
        type: integer
      file_type:
        $ref: '#/definitions/models.FileType'
      id:
        type: string
      is_processed:
        type: boolean
      metadata:
        type: string
      mime_type:
        type: string
      original_name:
        type: string
      page_count:
        type: integer
      updated_at:
        type: string
    type: object
  models.BookStatus:
    enum:
    - draft
    - published
    - archived
    type: string
    x-enum-varnames:
    - BookStatusDraft
    - BookStatusPublished
    - BookStatusArchived
  models.Bookmark:
    properties:
      book_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      label:
        type: string
      location:
        description: Может быть номером страницы, CFI для EPUB, и т.д.
        type: string
      user_id:
        type: string
    type: object
  models.Category:
    properties:
      children:
        items:
          $ref: '#/definitions/models.Category'
        type: array
      color:
        type: string
      created_at:
        type: string
      description:
        type: string
      icon:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      parent:
        $ref: '#/definitions/models.Category'
      parent_id:
        type: string
      slug:
        type: string
      sort_order:
        type: integer
      updated_at:
        type: string
    required:
    - name
    type: object
...
  models.CreateBookDTO:
    properties:
      author:
        type: string
      category_ids:
        items:
          type: string
        type: array
      copies_count:
        minimum: 0
        type: integer
      cover_url:
        type: string
      description:
        type: string
      is_premium:
        type: boolean
      isbn:
        type: string
      language:
        type: string
      page_count:
        type: integer
      publication_year:
        maximum: 9999
        minimum: 0
        type: integer
      publisher:
        type: string
      title:
        type: string
    required:
    - author
    - title
    type: object
  models.CreateBookmarkDTO:
    properties:
      book_id:
        type: string
      label:
        type: string
      location:
        type: string
    required:
    - book_id
    - location
    type: object
...
  models.ErrorResponseDTO:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
  models.FileType:
    enum:
    - pdf
    - epub
    - mobi
    type: string
    x-enum-varnames:
    - FileTypePDF
    - FileTypeEPUB
    - FileTypeMOBI
...
  models.SuccessResponseDTO:
    properties:
      data: {}
      message:
        type: string
    type: object
  models.UpdateBookDTO:
    properties:
      author:
        type: string
      category_ids:
        items:
          type: string
        type: array
      copies_count:
        minimum: 0
        type: integer
      cover_url:
        type: string
      description:
        type: string
      is_premium:
        type: boolean
      isbn:
        type: string
      language:
        type: string
      page_count:
        type: integer
      publication_year:
        maximum: 9999
        minimum: 0
        type: integer
      publisher:
        type: string
      status:
        $ref: '#/definitions/models.BookStatus'
      title:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for a digital library application, providing functionalities for
    book management, user authentication, and reading features.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Digital Library API
  version: "2.0"
paths:
...
  /bookmarks:
    post:
      consumes:
      - application/json
      description: Adds a new bookmark for a book by the authenticated user.
      parameters:
      - description: Bookmark data
        in: body
        name: bookmark
        required: true
        schema:
          $ref: '#/definitions/models.CreateBookmarkDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Bookmark'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponseDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponseDTO'
      security:
      - BearerAuth: []
      summary: Create a new bookmark
      tags:
      - Bookmarks
  /bookmarks/{id}:
    delete:
      description: Deletes a bookmark by its ID.
      parameters:
      - description: Bookmark ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponseDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponseDTO'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponseDTO'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponseDTO'
      security:
      - BearerAuth: []
      summary: Delete a bookmark
      tags:
      - Bookmarks
  /bookmarks/book/{book_id}:
    get:
      description: Retrieves all bookmarks for a specific book by the authenticated
        user.
      parameters:
      - description: Book ID
        in: path
        name: book_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Bookmark'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponseDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponseDTO'
      security:
      - BearerAuth: []
      summary: Get bookmarks for a book
      tags:
      - Bookmarks
  /books:
    get:
      description: Retrieves a paginated list of all available books.
      parameters:
      - default: 20
        description: Number of books to return
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.SuccessResponseDTO'
            - properties:
                Data:
                  items:
                    $ref: '#/definitions/models.Book'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponseDTO'
      summary: Get all books
      tags:
      - Books
    post:
      consumes:
      - application/json
      description: Adds a new book to the library. Librarian access required.
      parameters:
      - description: Book creation data
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/models.CreateBookDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/models.SuccessResponseDTO'
            - properties:
                Data:
                  $ref: '#/definitions/models.Book'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponseDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponseDTO'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.ErrorResponseDTO'
      security:
      - BearerAuth: []
      summary: Create a new book
      tags:
      - Books
  /books/{id}:
    delete:
      description: Deletes a book from the library. Librarian access required.
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponseDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponseDTO'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponseDTO'
      security:
      - BearerAuth: []
      summary: Delete a book
      tags:
      - Books
    get:
      description: Retrieves the details of a specific book by its UUID.
      parameters:
      - description: Book ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Book'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponseDTO'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponseDTO'
      summary: Get a book by ID
      tags:
      - Books
    put:
      consumes:
      - application/json
      description: Updates the details of a specific book. Librarian access required.
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      - description: Book data to update
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/models.UpdateBookDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.SuccessResponseDTO'
            - properties:
                Data:
                  $ref: '#/definitions/models.Book'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponseDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponseDTO'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponseDTO'
      security:
      - BearerAuth: []
      summary: Update a book
      tags:
      - Books
  /books/{id}/recommendations:
    get:
      description: Retrieves a list of recommended books based on the reading history
        of users who read the specified book.
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      - default: 10
        description: Limit for recommendations
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.SuccessResponseDTO'
            - properties:
                Data:
                  items:
                    $ref: '#/definitions/models.Book'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponseDTO'
      summary: Get book recommendations
      tags:
      - Books
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and a JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"