# internal/handlers/book_handler.go
package handlers

import (
	"github.com/oneErrortime/afst/internal/models"
	"github.com/oneErrortime/afst/internal/services"
	"net/http"
	"strconv"

	"github.com/gin-gonic/gin"
	"github.com/go-playground/validator/v10"
	"github.com/google/uuid"
)

// BookHandler обрабатывает запросы для книг
type BookHandler struct {
	bookService services.BookService
	validator   *validator.Validate
}

// NewBookHandler создает новый экземпляр BookHandler
func NewBookHandler(bookService services.BookService, validator *validator.Validate) *BookHandler {
	return &BookHandler{
		bookService: bookService,
		validator:   validator,
	}
}

// CreateBook godoc
// @Summary		Create a new book
// @Description	Adds a new book to the library. Librarian access required.
// @Tags			Books
// @Accept			json
// @Produce		json
// @Security		BearerAuth
// @Param			book	body		models.CreateBookDTO	true	"Book creation data"
// @Success		201		{object}	models.SuccessResponseDTO{Data=models.Book}
// @Failure		400		{object}	models.ErrorResponseDTO
// @Failure		401		{object}	models.ErrorResponseDTO
// @Failure		409		{object}	models.ErrorResponseDTO
// @Router			/books [post]
func (h *BookHandler) CreateBook(c *gin.Context) {
	var req models.CreateBookDTO

	// Парсим JSON
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, models.ErrorResponseDTO{
			Error:   "Неверный формат данных",
			Message: err.Error(),
		})
		return
	}

	// Валидируем данные
	if err := h.validator.Struct(&req); err != nil {
		c.JSON(http.StatusBadRequest, models.ErrorResponseDTO{
			Error:   "Ошибка валидации",
			Message: err.Error(),
		})
		return
	}

	// Вызываем сервис
	book, err := h.bookService.CreateBook(&req)
	if err != nil {
		c.JSON(http.StatusConflict, models.ErrorResponseDTO{
			Error:   "Ошибка создания книги",
			Message: err.Error(),
		})
		return
	}

	c.JSON(http.StatusCreated, models.SuccessResponseDTO{
		Message: "Книга создана успешно",
		Data:    book,
	})
}

// GetBook godoc
// @Summary		Get a book by ID
// @Description	Retrieves the details of a specific book by its UUID.
// @Tags			Books
// @Produce		json
// @Param			id	path		string	true	"Book ID"	Format(uuid)
// @Success		200	{object}	models.Book
// @Failure		400	{object}	models.ErrorResponseDTO
// @Failure		404	{object}	models.ErrorResponseDTO
// @Router			/books/{id} [get]
func (h *BookHandler) GetBook(c *gin.Context) {
	idParam := c.Param("id")
	id, err := uuid.Parse(idParam)
	if err != nil {
		c.JSON(http.StatusBadRequest, models.ErrorResponseDTO{
			Error:   "Неверный ID книги",
			Message: "ID должен быть в формате UUID",
		})
		return
	}

	book, err := h.bookService.GetBookByID(id)
	if err != nil {
		c.JSON(http.StatusNotFound, models.ErrorResponseDTO{
			Error:   "Книга не найдена",
			Message: err.Error(),
		})
		return
	}

	c.JSON(http.StatusOK, book)
}

// GetAllBooks godoc
// @Summary		Get all books
// @Description	Retrieves a paginated list of all available books.
// @Tags			Books
// @Produce		json
// @Param			limit	query		int	false	"Number of books to return"				default(20)
// @Param			offset	query		int	false	"Offset for pagination"					default(0)
// @Success		200		{object}	models.SuccessResponseDTO{Data=[]models.Book}
// @Failure		500		{object}	models.ErrorResponseDTO
// @Router			/books [get]
func (h *BookHandler) GetAllBooks(c *gin.Context) {
	// Парсим параметры пагинации
	limitParam := c.DefaultQuery("limit", "20")
	offsetParam := c.DefaultQuery("offset", "0")

	limit, err := strconv.Atoi(limitParam)
	if err != nil || limit <= 0 {
		limit = 20
	}

	offset, err := strconv.Atoi(offsetParam)
	if err != nil || offset < 0 {
		offset = 0
	}

	books, err := h.bookService.GetAllBooks(limit, offset)
	if err != nil {
		c.JSON(http.StatusInternalServerError, models.ErrorResponseDTO{
			Error:   "Ошибка получения книг",
			Message: err.Error(),
		})
		return
	}

	c.JSON(http.StatusOK, models.SuccessResponseDTO{
		Message: "Книги получены успешно",
		Data:    books,
	})
}

// UpdateBook godoc
// @Summary		Update a book
// @Description	Updates the details of a specific book. Librarian access required.
// @Tags			Books
// @Accept			json
// @Produce		json
// @Security		BearerAuth
// @Param			id		path		string					true	"Book ID"
// @Param			book	body		models.UpdateBookDTO	true	"Book data to update"
// @Success		200		{object}	models.SuccessResponseDTO{Data=models.Book}
// @Failure		400		{object}	models.ErrorResponseDTO
// @Failure		401		{object}	models.ErrorResponseDTO
// @Failure		404		{object}	models.ErrorResponseDTO
// @Router			/books/{id} [put]
func (h *BookHandler) UpdateBook(c *gin.Context) {
	idParam := c.Param("id")
	id, err := uuid.Parse(idParam)
	if err != nil {
		c.JSON(http.StatusBadRequest, models.ErrorResponseDTO{
			Error:   "Неверный ID книги",
			Message: "ID должен быть в формате UUID",
		})
		return
	}

	var req models.UpdateBookDTO

	// Парсим JSON
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, models.ErrorResponseDTO{
			Error:   "Неверный формат данных",
			Message: err.Error(),
		})
		return
	}

	// Валидируем данные
	if err := h.validator.Struct(&req); err != nil {
		c.JSON(http.StatusBadRequest, models.ErrorResponseDTO{
			Error:   "Ошибка валидации",
			Message: err.Error(),
		})
		return
	}

	// Вызываем сервис
	book, err := h.bookService.UpdateBook(id, &req)
	if err != nil {
		c.JSON(http.StatusNotFound, models.ErrorResponseDTO{
			Error:   "Ошибка обновления книги",
			Message: err.Error(),
		})
		return
	}

	c.JSON(http.StatusOK, models.SuccessResponseDTO{
		Message: "Книга обновлена успешно",
		Data:    book,
	})
}

// DeleteBook godoc
// @Summary		Delete a book
// @Description	Deletes a book from the library. Librarian access required.
// @Tags			Books
// @Produce		json
// @Security		BearerAuth
// @Param			id	path		string	true	"Book ID"
// @Success		200	{object}	models.SuccessResponseDTO
// @Failure		400	{object}	models.ErrorResponseDTO
// @Failure		401	{object}	models.ErrorResponseDTO
// @Failure		404	{object}	models.ErrorResponseDTO
// @Router			/books/{id} [delete]
func (h *BookHandler) DeleteBook(c *gin.Context) {
	idParam := c.Param("id")
	id, err := uuid.Parse(idParam)
	if err != nil {
		c.JSON(http.StatusBadRequest, models.ErrorResponseDTO{
			Error:   "Неверный ID книги",
			Message: "ID должен быть в формате UUID",
		})
		return
	}

	err = h.bookService.DeleteBook(id)
	if err != nil {
		c.JSON(http.StatusNotFound, models.ErrorResponseDTO{
			Error:   "Ошибка удаления книги",
			Message: err.Error(),
		})
		return
	}

	c.JSON(http.StatusOK, models.SuccessResponseDTO{
		Message: "Книга удалена успешно",
	})
}

// GetRecommendations godoc
// @Summary		Get book recommendations
// @Description	Retrieves a list of recommended books based on the reading history of users who read the specified book.
// @Tags			Books
// @Produce		json
// @Param			id		path		string	true	"Book ID"
// @Param			limit	query		int		false	"Limit for recommendations"	default(10)
// @Success		200		{object}	models.SuccessResponseDTO{Data=[]models.Book}
// @Failure		400		{object}	models.ErrorResponseDTO
// @Failure		500		{object}	models.ErrorResponseDTO
// @Router			/books/{id}/recommendations [get]
func (h *BookHandler) GetRecommendations(c *gin.Context) {
	idParam := c.Param("id")
	id, err := uuid.Parse(idParam)
	if err != nil {
		c.JSON(http.StatusBadRequest, models.ErrorResponseDTO{
			Error:   "Неверный ID книги",
			Message: "ID должен быть в формате UUID",
		})
		return
	}

	limit, _ := strconv.Atoi(c.DefaultQuery("limit", "10"))

	books, err := h.bookService.GetRecommendations(id, limit)
	if err != nil {
		c.JSON(http.StatusInternalServerError, models.ErrorResponseDTO{
			Error:   "Ошибка получения рекомендаций",
			Message: err.Error(),
		})
		return
	}

	c.JSON(http.StatusOK, models.SuccessResponseDTO{
		Message: "Рекомендации успешно получены",
		Data:    books,
	})
}